Se guardan y se obtienen de la BD los resaltadores y anotaciones por libro y por usuario
