--- Producto/Código Fuente/Baku/Backend/src/routes/marcador.router.ts
@@ -1,10 +1,0 @@
-import { Router } from 'express'
-import * as marcadorCtrl from './marcador.controller'
-
-const router = Router();
-
-router.post('/guardarNota', marcadorCtrl.guardarNota);
-
-router.get('/marcador', marcadorCtrl.getMarcador);
-
-export default router
